+++ src/rust/src/r_function/r_linalg.rs
@@ -109,6 +109,33 @@ fn rs_cov2cor(x: RMatrix<f64>) -> extendr_api::RArray<f64, [usize; 2]> {
     faer_to_r_matrix(cor.as_ref())
 }
 
+/// Calculates the mutual information matrix
+///
+/// @description Calculates the mutual information across all columns in the
+/// data.
+///
+/// @param x R matrix with doubles for which to calculate the mutual information
+/// @param n_bins Optional integer. Number of bins to use. If `NULL` is provided
+/// the function will default to `sqrt(nrows(x))`.
+/// @param normalise Boolean. Shall the normalised mutual information be
+/// calculated via joint entropy.
+///
+/// @returns The mutual information matrix.
+///
+/// @export
+#[extendr]
+fn rs_mutual_info(
+    x: RMatrix<f64>,
+    n_bins: Option<usize>,
+    normalise: bool,
+) -> extendr_api::RArray<f64, [usize; 2]> {
+    let mat = r_matrix_to_faer(&x);
+
+    let mi_mat = column_mutual_information(&mat, n_bins, normalise);
+
+    faer_to_r_matrix(mi_mat.as_ref())
+}
+
 /// Calculate the column wise correlations.
 ///
 /// @description Calculates the correlation matrix of the columns. This function
@@ -381,6 +408,7 @@ extendr_module! {
   fn rs_cos;
   fn rs_cor2;
   fn rs_cov2cor;
+  fn rs_mutual_info;
   fn rs_prcomp;
   fn rs_random_svd;
   fn rs_cor_upper_triangle;